mvp-backend/requirements.txt
requests
beautifulsoup4
python-telegram-bot
supabase-py
schedule